requirements.txt
asgiref==3.2.10
Django==3.1
django-cleanup==5.0.0
django-recaptcha==2.0.6
numpy==1.19.1
opencv-python==4.4.0.42
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
